// src/app/core/services/models/page-response-mission-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { MissionResponse } from '../models/mission-response';
export interface PageResponseMissionResponse {
  content?: Array<MissionResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}